+++ NEWS.md
@@ -12,6 +12,10 @@
   
   - `slide_max()`, `slide_index_max()`: for rolling maximums
 
+* `slide_index_*()` functions are now a bit faster from using
+  `vctrs::vec_unrep()` to locate unique index values, rather than
+  `vctrs::vec_group_loc()`.
+
 * `vignette("rowwise")` has been updated to use `cur_data()` from dplyr 1.0.0,
   which makes it significantly easier to do rolling operations on data frames
   (like rolling regressions) using slider in a dplyr pipeline.